Update modal dialog text clarity
Change the text in the modal dialog to make the confirmation messages clearer. "Yes, Clear it!" and "Clear Code?" were switched to make more sense with the actions taken. The confirmation request was also updated from "Are you sure you want to clear current code?" to "Are you sure you want to clear the current code?" for grammatical accuracy.