Improve website UI and IDE setup
Resolved overflow issues in CSS and enhanced logo alignment in the navbar for a better UI experience. The CSS 'overflow: auto' property and a fixed height have been added to control overflow, improving the user experience when content is too big for its design. Additionally, the orientation of the logo image and text in the navbar has been updated to 'vertical-align: middle', which helps them integrate better with other elements.

On the IDE side, libraries for "halfmoon", "bootstrap", and "monaco-editor" have been added to the project for smoother development interface and better code management.